requirements.txt
Flask==2.3.3
openpyxl==3.1.2
boto3==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0
